// frontend/src/api/model/bedStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Hospital Internment API
 * OpenAPI spec version: 0.1.0
 */

export type BedStatus = typeof BedStatus[keyof typeof BedStatus];


export const BedStatus = {
  AVAILABLE: 'AVAILABLE',
  RESERVED: 'RESERVED',
  OCCUPIED: 'OCCUPIED',
  PENDING_CLEANING: 'PENDING_CLEANING',
  BLOCKED: 'BLOCKED',
  MAINTENANCE: 'MAINTENANCE',
} as const;